style(types): extract all to seperate files
