+++ prettier.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  trailingComma: 'es5',
+  trailingComma: 'all',
   tabWidth: 2,
   semi: true,
   singleQuote: true,
