+++ src/constants.ts
@@ -1,0 +1,8 @@
+export const API_BASE_URL = 'https://api.3commas.io';
+
+export const API_BASE_PATHS = {
+  1: '/public/api/ver1',
+  2: '/public/api/v2',
+};
+
+export const WEBSOCKET_URL = 'wss://ws.3commas.io/websocket';
