+++ src/interfaces/bots-stats-params.interface.ts
@@ -1,0 +1,4 @@
+export interface BotsStatsParams {
+  account_id?: number;
+  bot_id?: number;
+}
