+++ src/interfaces/currency-params.interface.ts
@@ -1,0 +1,4 @@
+export interface CurrencyParams {
+  market_code?: string;
+  pair: string;
+}
