+++ src/interfaces/fund-params.interface.ts
@@ -1,0 +1,9 @@
+export interface FundParams {
+  order_type: 'market' | 'limit';
+  units: {
+    value: number | string;
+  };
+  price?: {
+    value: number | string;
+  };
+}
