+++ src/interfaces/index.ts
@@ -1,0 +1,23 @@
+export * from './api-options.interface';
+export * from './bots-params.interface';
+export * from './bots-stats-params.interface';
+export * from './currency-params.interface';
+export * from './deals-params.interface';
+export * from './fund-params.interface';
+export * from './leverage.interface';
+export * from './market-currency-params.interface';
+export * from './position-conditional-price.interface';
+export * from './position-conditional.interface';
+export * from './position.interface';
+export * from './smart-trade-history-params.interface';
+export * from './smart-trade-params.interface';
+export * from './step-price.interface';
+export * from './step.interface';
+export * from './stop-loss.interface';
+export * from './take-profit.interface';
+export * from './three-commas-error.interface';
+export * from './timeout.interface';
+export * from './trailing.interface';
+export * from './transfer-history-params.interface';
+export * from './transfer-params.interface';
+export * from './units-class.interface';
