+++ src/interfaces/leverage.interface.ts
@@ -1,0 +1,5 @@
+export interface Leverage {
+  enabled: boolean;
+  type?: 'custom' | 'cross';
+  value?: number;
+}
