+++ src/interfaces/market-currency-params.interface.ts
@@ -1,0 +1,4 @@
+export interface MarketCurrencyParams {
+  market_code: string;
+  pair: string;
+}
