+++ src/interfaces/position-conditional-price.interface.ts
@@ -1,0 +1,4 @@
+export interface PositionConditionalPrice {
+  value: number;
+  type: 'bid' | 'ask' | 'last';
+}
