+++ src/interfaces/step-price.interface.ts
@@ -1,0 +1,5 @@
+export interface StepPrice {
+  value?: number;
+  type: 'bid' | 'ask' | 'last';
+  percent?: number;
+}
