+++ src/interfaces/stop-loss-conditional.ts
@@ -1,0 +1,7 @@
+import { StepPrice } from './step-price.interface';
+import { Trailing } from './trailing.interface';
+
+export interface StopLossConditional {
+  price: StepPrice;
+  trailing?: Trailing;
+}
