+++ src/interfaces/take-profit.interface.ts
@@ -1,0 +1,6 @@
+import { Step } from './step.interface';
+
+export interface TakeProfit {
+  enabled: boolean;
+  steps?: Step[];
+}
