+++ src/interfaces/three-commas-error.interface.ts
@@ -1,0 +1,7 @@
+export interface ThreeCommasError {
+  error: string;
+  error_description?: string;
+  error_attributes?: {
+    [key: string]: string;
+  };
+}
