+++ src/interfaces/timeout.interface.ts
@@ -1,0 +1,4 @@
+export interface Timeout {
+  enabled: boolean;
+  value: number;
+}
