+++ src/interfaces/trailing.interface.ts
@@ -1,0 +1,4 @@
+export interface Trailing {
+  enabled: boolean;
+  percent?: number;
+}
