+++ src/interfaces/transfer-history-params.interface.ts
@@ -1,0 +1,6 @@
+export interface TransferHistoryParams {
+  account_id: number | string;
+  currency: string;
+  page?: number | string;
+  per_page?: number | string;
+}
