+++ src/interfaces/transfer-params.interface.ts
@@ -1,0 +1,6 @@
+export interface TransferParams {
+  currency: string;
+  amount: number | string;
+  from_account_id: number | string;
+  to_account_id: number | string;
+}
