+++ src/interfaces/units-class.interface.ts
@@ -1,0 +1,3 @@
+export interface UnitsClass {
+  value: number;
+}
