+++ src/types/channel.type.ts
@@ -1,0 +1,1 @@
+export type Channel = 'SmartTradesChannel' | 'DealsChannel';
