+++ src/types/index.ts
@@ -1,0 +1,2 @@
+export * from './channel.type';
+export * from './websocket-callback.type';
