+++ src/types/websocket-callback.type.ts
@@ -1,0 +1,3 @@
+import { Data } from 'ws';
+
+export type WebsocketCallback = (data: Data) => void;
